% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_org2}
\alias{data_org2}
\title{Large list with 2 elements: TrainLong, ValidLongs}
\format{
An object of class \code{list} of length 2.
}
\usage{
data_org2
}
\description{
Large list with 2 elements: TrainLong, ValidLongs
}
\keyword{datasets}
