% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{follow_up_time}
\alias{follow_up_time}
\title{Sample Data of the MASTA Pacakge}
\format{A data.frame with 3 columns:
\describe{
  \item{id}{patient id.}
  \item{fu_time}{the follow up time for the patient.}
  \item{train_valid}{indicator for patients in the training data set; 1=training, 0=test}
}}
\usage{
follow_up_time
}
\description{
The follow up time of training and validation data for illustrating the use of the MASTA algorithm
}
\keyword{datasets}
