% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA_related.R
\name{fpca.combine}
\alias{fpca.combine}
\title{Functional Principal Component Analysis (FPCA)}
\usage{
fpca.combine(
  time_code,
  follow_up_time,
  K.select = "PropVar",
  n.grid = 401,
  propvar = 0.85
)
}
\arguments{
\item{time_code}{longitudinal encounter times. They should be greater than or equal to 1.}

\item{follow_up_time}{follow-up time (training)}

\item{K.select}{characters indicating which method to choose the number of principal components K.
Default is K.select="PropVar", and K.select="PPIC" is also available.}

\item{n.grid}{an integer value for grid points used in estimating covariance function g. Default is \code{401}.}

\item{propvar}{a proportion of variation used to select number of FPCs. Default is \code{0.85}.}
}
\description{
Performs FPCA to extract features from longitudinal encounter data.
}
