% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longitudinal}
\alias{longitudinal}
\title{Sample Data of the MASTA Pacakge}
\format{A data.frame with 3 columns:
\describe{
  \item{code}{integer indicator for the type of codes. It must be incremental and start from 1, which contains no skipping number.}
  \item{id}{patient id.}
  \item{time}{encounter time with the code. One patient can encounter with different types of codes multiple times.}
}}
\usage{
longitudinal
}
\description{
A long form longitudinal data for illustrating the use of the MASTA algorithm
}
\keyword{datasets}
