% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTA.R
\docType{package}
\name{MASTA-package}
\alias{MASTA-package}
\alias{MASTA}
\title{MASTA Package}
\description{
Implements an algorithm to identify the occurrence of event outcome from trajectories of several predictors.
}
\details{
Visit the documentation website for more details. https://celehs.github.io/MASTA/
}
