% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masta-fit.R
\name{masta.fit}
\alias{masta.fit}
\title{Main function implementing the MASTA algorithm}
\usage{
masta.fit(
  object,
  survival,
  follow_up_time,
  Tend = 1,
  cov_group = NULL,
  thresh = 0.7,
  PCAthresh = 0.9,
  seed = 100
)
}
\arguments{
\item{object}{The object returned by the \code{fpca.combine} function}

\item{survival}{the labeled data. The columns should be 1) id, 2) event indicator, 3) event time, followed by baseline predictors.}

\item{follow_up_time}{the follow-up data}

\item{cov_group}{a vector of consecutive integers describing the grouping only for covariates. When \code{NULL} is specified (default), each covariate will be in different group.}

\item{thresh}{a default is \code{0.7}, which means if there are codes with >70\% patients no codes, only use first code time.}

\item{PCAthresh}{a threshold value for PCA. Default is \code{0.9}.}

\item{seed}{random seed used for the sampling. Default is \code{100}.}
}
\value{
A list with components:

\item{bgbbest_FromChengInit_BFGS}{Details of the fitted model}

\item{Cstat_BrierSc_ChengInit_BFGS}{Performance of the derived algorithm. C-statistics, etc.}

\item{group}{A vector of consecutive integers describing the grouping coefficients}
}
\description{
This function builds an algorithm to identify the occurrence of event outcome from trajectories of several predictors.
}
