% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masta-validation.R
\name{masta_validation}
\alias{masta_validation}
\title{Function for validating the MASTA algorithm with new data}
\usage{
masta_validation(object, new_longitudinal, new_follow_up_time, new_survival)
}
\arguments{
\item{object}{The object returned by the \code{masta.fit} function}

\item{new_longitudinal}{longitudinal encounter times for the validation. See an example as \code{longitudinal} data.}

\item{new_follow_up_time}{the follow-up data for the validation}

\item{new_survival}{the labeled data for the validation. The columns should be 1) id, 2) event indicator, 3) event time, followed by baseline predictors.}
}
\value{
A list with components:

\item{result_valid}{Performance of the derived algorithm. C-statistics, etc.}

\item{risk_score_valid}{Predcited Score}

\item{pred_surv_valid}{Predicted incidence curves}
}
\description{
This function builds an algorithm to identify the occurrence of event outcome from trajectories of several predictors.
}
