% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATA.R
\docType{package}
\name{MATA-package}
\alias{MATA-package}
\alias{MATA}
\title{MATA Package}
\description{
Implements an algorithm to identify the occurrence of event outcome from trajectories of several predictors.
}
\details{
Visit the documentation website for more details. https://celehs.github.io/MATA/
}
