% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_follow_up_time}
\alias{new_follow_up_time}
\title{Sample Data of the MASTA Pacakge}
\format{
A data.frame with 2 columns:
\describe{
  \item{id}{patient id.}
  \item{fu_time}{the follow up time for the patient.}
}
}
\usage{
new_follow_up_time
}
\description{
The follow up time of training and validation data for illustrating how to validate the MASTA algorithm with a new data
}
\keyword{datasets}
