% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_survival}
\alias{new_survival}
\title{Sample Data of the MASTA Pacakge}
\format{
A data.frame with 6 columns:
\describe{
  \item{id}{patient id.}
  \item{event_ind}{indicator for events; 1=event, 0=cencoring}
  \item{event_time}{the event time if \code{event_ind}=1.}
  \item{cov_1,cov_2,cov_3}{3 covariates used in a survival regression model.}
}
}
\usage{
new_survival
}
\description{
Survival data for illustrating how to validate the MASTA algorithm with a new data
}
\keyword{datasets}
