% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETLER-package.R
\docType{package}
\name{PETLER-package}
\alias{PETLER-package}
\alias{PETLER}
\title{PETLER Package}
\description{
Implements an algorithm to identify the occurrence of event outcome from trajectories of several predictors.
}
\references{
Wu, S., Müller, H., & Zhang, Z. (2013). FUNCTIONAL DATA ANALYSIS FOR POINT PROCESSES WITH RARE EVENTS. Statistica Sinica, 23(1), 1-23.
}
