% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETLER-package.R
\name{petler}
\alias{petler}
\title{Main function implementing the PETLER algorithm}
\usage{
petler(
  object,
  cov_group = NULL,
  thresh = 0.7,
  PCAthresh = 0.9,
  seed = 1234,
  seed2 = 100
)
}
\arguments{
\item{object}{...}

\item{cov_group}{...}

\item{thresh}{...}

\item{PCAthresh}{...}

\item{seed}{...}

\item{seed2}{...}
}
\value{
A list with components:

\item{bgbbest_FromChengInit_BFGS}{Details of the fitted model}

\item{Cstat_BrierSc_ChengInit_BFGS}{Performance of the derived algorithm. C-statistics, etc.}

\item{group}{A vector of consecutive integers describing the grouping coefficients}

Several output files will be saved in the \code{outdir} directory.
}
\description{
...
}
