% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z}
\alias{Z}
\title{Z List with 13 elements: TrainFt, ValidFt, TrainPK, ValidPK, TrainN, ValidN, nn, codes, Tend, TrainSurv, ValidSurv, TrainSurv_pred_org, ValidSurv_pred_org}
\format{
An object of class \code{list} of length 13.
}
\usage{
Z
}
\description{
Z List with 13 elements: TrainFt, ValidFt, TrainPK, ValidPK, TrainN, ValidN, nn, codes, Tend, TrainSurv, ValidSurv, TrainSurv_pred_org, ValidSurv_pred_org
}
\keyword{datasets}
