% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTA.R
\docType{data}
\name{follow_up_time}
\alias{follow_up_time}
\title{Sample Data of the MASTA Pacakge}
\format{
A list with 4 data tables:
\describe{
  \item{TrainSurv}{baseline survival data for training (labeled); 1st colum: patient id, 2nd colum: event indicator (1=event, 0=censoring), 3rd colum: event time, 4th colum: follow-up time, 5th colum--: covariates.}
  \item{ValidSurv}{baseline survival data for validation; baseline survival data for training (labeled); 1st colum: patient id, 2nd colum: event indicator (1=event, 0=censoring), 3rd colum: event time, 4th colum: follow-up time, 5th colum--: covariates.}
  \item{TrainCode}{1st colum: patient id, 2nd colum: follow-up time, 3rd colum: time label (month), 4th colum-- : predictors.}
  \item{ValidCode}{1st colum: patient id, 2nd colum: follow-up time, 3rd colum: time label (month), 4th colum-- : predictors.}
}
}
\usage{
follow_up_time
}
\description{
A list of training and validation data for illustrating the use of the MASTA algorithm
}
\keyword{datasets}
