% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETLER.R
\docType{package}
\name{PETLER-package}
\alias{PETLER-package}
\alias{PETLER}
\title{PETLER Package}
\description{
Implements an algorithm to identify the occurrence of event outcome from trajectories of several predictors.
}
\details{
Visit the documentation website for more details. https://celehs.github.io/PETLER/
}
