% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETLER-package.R
\name{petler.base}
\alias{petler.base}
\title{...}
\usage{
petler.base(data, PPIC_K = FALSE, n.grid = 401, propvar = 0.85, n_core = 4)
}
\arguments{
\item{data}{...}

\item{PPIC_K}{...}

\item{n.grid}{...}

\item{propvar}{...}

\item{n_core}{...}
}
\description{
...
}
